better structure with import and export
