+++ Html/booking.html
@@ -19,7 +19,7 @@
         /></a>
         <h1>Lule Northern Lights Cinema</h1>
         <img
-          onclick="hamburger()"
+        
           id="hamburger-icon"
           src="../static/hamburger-icon.svg"
           alt="hamburger-icon"
@@ -162,7 +162,7 @@
     </footer>
 
     <script src="../js/booking/booking.js"></script>
-    <script src="/js/script.js"></script>
+    <script type="module" src="/js/script.js"></script>
 
     <script type="module">
       import renderSecNav from "../js/secNav/secNav.js"
