+++ Html/upcoming.html
@@ -119,7 +119,7 @@
             <img id="cinema-collab" src="../static/europa-cinemas.png" alt="cinema collab" />
         </div>
     </footer>
-    <script src="./js/script.js"></script>
+    <script type="module" src="/js/script.js"></script>
     <script src="../js/secNav/secNav.js" type="module"></script>
 </body>
