+++ Html/wholeProgramPage.html
@@ -11,7 +11,6 @@
   <body>
     <header class="header">
       <div class="top-header">
-        
         <a href="/index.html">
           <img
             class="logotype"
@@ -19,7 +18,7 @@
             alt="logotype for the cinema"
         /></a>
         <h1>Lule Northern Lights Cinema</h1>
-        <img onclick="hamburger()"
+        <img
           id="hamburger-icon"
           src="../static/hamburger-icon.svg"
           alt="hamburger-icon"
@@ -108,7 +107,6 @@
     <main>
       <ul id="ul"></ul>
     </main>
-   
 
     <footer class="footer">
       <div class="footer-sections">
@@ -158,14 +156,12 @@
         />
       </div>
     </footer>
-    <script src="/js/script.js"></script>
+    <script type="module" src="/js/script.js"></script>
     <script src="../js/movieRows.js" type="module"></script>
-    
+
     <script type="module">
-        import renderSecNav from "../js/secNav/secNav.js"
-        renderSecNav(""); //Pass along emtpy folderDir since the secNav html-files are in the same directory as this one
-      </script>
-
-</body>
-
+      import renderSecNav from "../js/secNav/secNav.js";
+      renderSecNav(""); //Pass along emtpy folderDir since the secNav html-files are in the same directory as this one
+    </script>
+  </body>
 </html>
