+++ html/bistro-menu.html
@@ -167,7 +167,7 @@
         />
       </div>
     </footer>
-    <script src="./js/script.js"></script>
+    <script type="module" src="/js/script.js"></script>
     <script type="module">
       import renderSecNav from "../js/secNav/secNav.js";
       renderSecNav(""); //Pass along emtpy folderDir since the secNav html-files are in the same directory as this one
