+++ index.html
@@ -19,7 +19,6 @@
         /></a>
         <h1>Lule Northern Lights Cinema</h1>
         <img
-          onclick="hamburger()"
           id="hamburger-icon"
           src="../static/hamburger-icon.svg"
           alt="hamburger-icon"
@@ -159,7 +158,12 @@
     </footer>
     <script type="text/javascript" src="/src/movies.json"></script>
 
-    <script src="./js/script.js"></script>
+    <script type="module" src="./js/script.js"></script>
+
+    <script type="module">
+      import { displayCards } from "/js/movieCards/movieCards.js";
+      displayCards();
+    </script>
 
     <script type="module">
       import renderSecNav from "/js/secNav/secNav.js";
