+++ js/movieRows.js
@@ -1,38 +1,32 @@
 const ul = document.getElementById("ul");
 
 async function movieRows() {
+  const res = await fetch("../src/movies.json");
+  const data = await res.json();
 
-    const res = await fetch("../src/movies.json");
-    const data = await res.json();
-    
-    
-    data.movies.forEach(element => { 
-        
-        const li = document.createElement("li");
-        const div = document.createElement("div");
+  data.movies.forEach((element) => {
+    const li = document.createElement("li");
+    const div = document.createElement("div");
 
-        const img = document.createElement("img");
-        img.src = element.posterImageUrl;
-        
-        const title = document.createElement("h2");
-        title.innerHTML = element.title;
+    const img = document.createElement("img");
+    img.src = element.posterImageUrl;
 
-        const tickets = document.createElement("a");
-        tickets.innerText = "Biljetter";
-        tickets.href = "#";
+    const title = document.createElement("h2");
+    title.innerHTML = element.title;
 
-        const moreInfo = document.createElement("a"); 
-        moreInfo.innerText = "Mer info";
-        moreInfo.href = "#";
+    const tickets = document.createElement("a");
+    tickets.innerText = "Biljetter";
+    tickets.href = "#";
 
-        div.append(title, tickets, moreInfo); 
-        li.append(img, div);
-        
-        ul.append(li);
-        
-    });
-        
+    const moreInfo = document.createElement("a");
+    moreInfo.innerText = "Mer info";
+    moreInfo.href = "#";
+
+    div.append(title, tickets, moreInfo);
+    li.append(img, div);
+
+    ul.append(li);
+  });
 }
 
 movieRows();
-
